Major refactoring and "auto" mode
Previously, we created the badkeys blocklist with the "createlists"
shell script, which was a quick and dirty hack.
This adds an "auto" mode that does the work of that script in a cleaner
and pure-python way. "createlists" will soon be removed.
Intermediate processing of the individual sub-blocklists is cached
in "~/.cache/blocklistmaker", avoiding to re-parse known lists.
This is a massive speedup.

Bug: https://github.com/badkeys/blocklistmaker/issues/2
Closes: https://github.com/badkeys/blocklistmaker/issues/3